Rename jsx files wrongly named js
