+++ packages/ndla-howto/src/StaticInfoComponents/components/FactBoxExample.jsx
