+++ packages/ndla-howto/src/StaticInfoComponents/components/SolutionExample.jsx
