+++ packages/ndla-howto/src/StaticInfoComponents/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import AudioPlayerExamples from './components/AudioExample';
-import FactboxExample from './components/FactboxExample';
+import FactBoxExample from './components/FactBoxExample';
 import SolutionExample from './components/SolutionExample';
 import Markdown from './components/Markdown';
 
@@ -17,7 +17,7 @@ export const stories = {
     body: [
       {
         type: 'component',
-        content: FactboxExample,
+        content: FactBoxExample,
       },
     ],
   },
